+++ src/DotPulsar/Internal/Consumer.cs
@@ -351,17 +351,6 @@ public sealed class Consumer<TMessage> : IConsumer<TMessage>
         await Task.WhenAll(seekTasks).ConfigureAwait(false);
     }
 
-    [Obsolete("GetLastMessageId is obsolete. Please use GetLastMessageIds instead.")]
-    public async ValueTask<MessageId> GetLastMessageId(CancellationToken cancellationToken)
-    {
-        await Guard(cancellationToken).ConfigureAwait(false);
-
-        if (!_isPartitionedTopic)
-            return await _subConsumers[_subConsumerIndex].GetLastMessageId(cancellationToken).ConfigureAwait(false);
-
-        throw new NotSupportedException("GetLastMessageId can't be used on partitioned topics. Please use GetLastMessageIds");
-    }
-
     public async ValueTask<IEnumerable<MessageId>> GetLastMessageIds(CancellationToken cancellationToken)
     {
         await Guard(cancellationToken).ConfigureAwait(false);